pyproject.toml
[tool.poetry]
name = "video processing application"
version = "0.1.0"
description = ""
authors = ["SirAbhi13"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
ffmpeg-python = "^0.2.0"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
